Rename scenario to generator
